Downgrade pinger log level
Tough decision, by the measure of if this is the only thing wrong, do we
want to interrupt a users game to notify of this? The answer is no, if
there is something else wrong, we would like want and get that info, the
'ping' being broken is likely one more thing that would be busted. We
still print the info message OTOH to keep track that this happened.