Upgrade 'fine' error messages to 'info'
Arguably these should be a 'warn', an error happening where values that
are supposed to be saved are not, is more significant than 'fine' level
of logging. Going to INFO for now so these are printed to console
without turning on debug logging, not going with WARN yet as that should
add a pop-up which is perhaps not necessary.